requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.3.0
gunicorn==22.0.0
Werkzeug==3.1.3
WTForms==3.1.2
psycopg2-binary==2.9.9
